Ensure dummy routers have an associated service
This suppresses a Traefik error during startup,
which says that all routers should have a service defined.

We create a dummy service here, which leads to 127.0.0.1.

This suppreses the error, but has the downside that hitting the URL
(e.g. https://SOME_ADDITIONAL_SSL_CERTIFICATE_DOMAIN)
causes a redirect loop. In the future, we should do better.
Serving some dummy or 404 page would be better than a redirect loop.